# src/main/resources/openapi/management.yml
openapi: 3.0.0
info:
  version: 2.0.0
  title: DXI
paths:
...
  /student:
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64

      operationId: getAllStudents
      description: Returns all students
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentFindResponse'
  /student/{semesterId}:
    parameters:
      - name: semesterId
        in: path
        schema:
          type: integer
          format: int64
        required: true

    post:
      operationId: createStudent
      description: Create user Student
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64
          required: false

      operationId: getStudentsForSemester
      description: Gets users for semester
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentFindResponse'
...
  /student/{id}:
    put:
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      operationId: updateStudent
      description: Updates user student
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdate'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
    delete:
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      operationId: deleteStudent
      description: Deletes user Student
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /user:
    get:
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            format: int64
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            format: int64
          required: false
      operationId: getAllUsers
      description: Returns all users
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFindResponse'
  /user/{id}:
    delete:
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      operationId: delteUser
      description: Returns all users
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        email:
          type: string
        firstName:
          type: string
        isActive:
          type: boolean
    UserFindResponse:
      type: object
      properties:
        startElement:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
...
    Student:
      type: object
      properties:
        username:
          type: string
        id:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
    StudentFindResponse:
      type: object
      properties:
        startElement:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
    StudentUpdate:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
    StudentResponse:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        isActive:
          type: boolean
...
    Role:
      type: string
      enum:
        - ADMIN
        - LECTURER
        - STUDENT
        - USER